pyproject.toml
[project]
name = "bakeoff"
version = "0.0.1"

[tool.poetry]
name = "bakeoff"
version = "0.0.1"
description = ""
authors = ["Kevin Chan"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "1.87.0"
semantic-kernel = { version = "1.33.0", extras = ["mcp"] }
azure-identity = "1.17.1"
ollama = "^0.5.1"
chainlit = "^2.5.5"
langgraph = "^0.4.8"
langchain-core = "^0.3.66"
langchain-openai = "^0.3.25"
langchain-community = "^0.3.26"
agno = "^1.6.4"
openai-agents = "^0.0.19"

[tool.poetry.group.dev.dependencies]
ipykernel = "6.25.2"
jupyterlab = "4.0.6"
"keyrings.alt" = "5.0.0"
scikit-learn = "1.4.0"
pandas = "2.2.2"

[tool.poetry.group.ci.dependencies]
flake8-docstrings = "1.6.0"
flake8 = "4.0.1"
pep8-naming = "0.13.0"
pytest-cov = "3.0.0"
pytest-azurepipelines = "1.0.3"
pytest-mock = "3.7.0"
pytest = "8.3.2"
pytest-asyncio = "0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.workspace]
members = [
    "mcp-server",
]
